+++ src/test/PrimeNumberTest.java
@@ -1,5 +1,7 @@
+
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
@@ -16,15 +18,15 @@ class PrimeNumberTest {
 
     @Test
     void makePrimeNumber1() {
-        List<Integer> primeNumbers1 = primeNumber.makePrimeNumber(10);
+        List<Long> primeNumbers1 = primeNumber.makePrimeNumber(10);
         Assertions.assertEquals(4, primeNumbers1.size());
 
-        List<Integer> primeNumbers2 = primeNumber.makePrimeNumber(15);
+        List<Long> primeNumbers2 = primeNumber.makePrimeNumber(15);
         Assertions.assertEquals(6, primeNumbers2.size());
     }
 
     @Test
     void makePrimeNumber2() {
-        List<Integer> primeNumbers = primeNumber.makePrimeNumber(10000000);
+        List<Long> primeNumbers = primeNumber.makePrimeNumber(10000000);
     }
 }
